pyproject.toml
[project]
name = "contactsearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crawl4ai>=0.8.0",
    "fastapi[standard]>=0.128.0",
    "langchain>=1.2.8",
    "langchain-ollama>=1.0.1",
    "markdownify>=1.2.2",
    "ollama>=0.6.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
